% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dbi.R
\name{monitor_dbi}
\alias{monitor_dbi}
\title{Create a DBI Monitor}
\usage{
monitor_dbi(conn, auto_disconnect = FALSE, keep = FALSE)
}
\arguments{
\item{conn}{A \linkS4class{DBIConnection} object, as returned by
\code{\link[=dbConnect]{dbConnect()}}.}

\item{auto_disconnect}{whether the connection should be automatically closed
when the monitor is deleted. Set to \code{TRUE} if you initialised the
connection inside the call to \code{memory_dbi}.}

\item{keep}{whether to keep the tables created on exit. By default, tables
are removed.}
}
\value{
A \code{monitor} object.
}
\description{
Methods for creating database-backed (\pkg{DBI}-based) \code{monitor} objects.
}
