% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simmer.mon.R
\docType{package}
\name{simmer.mon-package}
\alias{simmer.mon-package}
\title{\pkg{simmer.mon}: Monitoring Backends for \pkg{simmer}}
\description{
Provides monitoring backends that can be attached to \code{simmer}
environments to replace the default in-memory monitor.
}
\seealso{
\pkg{simmer}'s homepage \url{http://r-simmer.org} and
GitHub repository \url{https://github.com/r-simmer/simmer.mon}.
}
\author{
Iñaki Ucar
}
